adds signup module
