+++ lib/auth/models/index.js
@@ -1,10 +1,7 @@
 'use strict';
 
-const express = require('express');
 const { Sequelize, DataTypes } = require('sequelize');
-const bcrypt = require('bcrypt');
-const base64 = require('base-64');
-const app = express();
+
 
 // Sequelize is a Constructor from the Sequelize library
 // DataType = object that lets you define data types to be stored in SQL - SQL has its own specific data types
