+++ lib/auth/models/user.schema.js
@@ -1,6 +1,6 @@
 'use strict';
 
-
+const bcrypt = require('bcrypt');
 
 let User = (sequelize, DataTypes) => sequelize.define('User', {
   username: {
@@ -12,5 +12,9 @@ let User = (sequelize, DataTypes) => sequelize.define('User', {
     allowNull: false,
   },
 });
+
+// User.beforeCreate(async user => {
+//   user.password = await bcrypt.hash(user.password, 10);
+// });
 
 module.exports = User;
