+++ lib/server.js
@@ -5,11 +5,11 @@ const express = require('express');
 // const { Sequelize, DataTypes } = require('sequelize');
 // const bcrypt = require('bcrypt');
 // const base64 = require('base-64');
-const signUpRoute = require('./routes/signUp.js');
+const userRoutes = require('./routes/userRoutes.js');
 const app = express();
 
 app.use(express.json());
-app.use(signUpRoute);
+app.use(userRoutes);
